Move HF example data prep to prepare script
